% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_access.R
\name{baato_API}
\alias{baato_API}
\title{Check if the json object returned by API call is valid}
\description{
Check if the json object returned by API call is valid

Check if the json object returned by API call is valid
}
\examples{

## ------------------------------------------------
## Method `baato_API$rev_api_call`
## ------------------------------------------------

baato$rev_api_call(lat = 27.70446921370009, lon = 85.32051086425783)

## ------------------------------------------------
## Method `baato_API$search_api_call`
## ------------------------------------------------

baato$search_api_call(q="shemrock nepalgunj",limit=4)

## ------------------------------------------------
## Method `baato_API$places_api_call`
## ------------------------------------------------

baato$places_api_call(placeId = 102235)

## ------------------------------------------------
## Method `baato_API$nearby_api_call`
## ------------------------------------------------

baato$nearby_api_call(type="school",lat=27.71765,lon=85.32691,limit=20)


## ------------------------------------------------
## Method `baato_API$styles_api_call`
## ------------------------------------------------

baato$styles_api_call(style_name="monochrome")

## ------------------------------------------------
## Method `baato_API$baato_directions_api`
## ------------------------------------------------

baato$baato_directions_api(points = point, mode="car")
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-baato_API-new}{\code{baato_API$new()}}
\item \href{#method-baato_API-set_key}{\code{baato_API$set_key()}}
\item \href{#method-baato_API-display_key}{\code{baato_API$display_key()}}
\item \href{#method-baato_API-validate_resp}{\code{baato_API$validate_resp()}}
\item \href{#method-baato_API-display_resp}{\code{baato_API$display_resp()}}
\item \href{#method-baato_API-rev_api_call}{\code{baato_API$rev_api_call()}}
\item \href{#method-baato_API-search_api_call}{\code{baato_API$search_api_call()}}
\item \href{#method-baato_API-places_api_call}{\code{baato_API$places_api_call()}}
\item \href{#method-baato_API-nearby_api_call}{\code{baato_API$nearby_api_call()}}
\item \href{#method-baato_API-styles_api_call}{\code{baato_API$styles_api_call()}}
\item \href{#method-baato_API-baato_directions_api}{\code{baato_API$baato_directions_api()}}
\item \href{#method-baato_API-clone}{\code{baato_API$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baato_API-new"></a>}}
\if{latex}{\out{\hypertarget{method-baato_API-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baato_API$new(key = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baato_API-set_key"></a>}}
\if{latex}{\out{\hypertarget{method-baato_API-set_key}{}}}
\subsection{Method \code{set_key()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baato_API$set_key(key)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baato_API-display_key"></a>}}
\if{latex}{\out{\hypertarget{method-baato_API-display_key}{}}}
\subsection{Method \code{display_key()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baato_API$display_key()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baato_API-validate_resp"></a>}}
\if{latex}{\out{\hypertarget{method-baato_API-validate_resp}{}}}
\subsection{Method \code{validate_resp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baato_API$validate_resp(req)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{req}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
error message
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baato_API-display_resp"></a>}}
\if{latex}{\out{\hypertarget{method-baato_API-display_resp}{}}}
\subsection{Method \code{display_resp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baato_API$display_resp(req)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{req}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the API response in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baato_API-rev_api_call"></a>}}
\if{latex}{\out{\hypertarget{method-baato_API-rev_api_call}{}}}
\subsection{Method \code{rev_api_call()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baato_API$rev_api_call(lat, lon)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lat}}{}

\item{\code{lon}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the API response in JSON format
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{baato$rev_api_call(lat = 27.70446921370009, lon = 85.32051086425783)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baato_API-search_api_call"></a>}}
\if{latex}{\out{\hypertarget{method-baato_API-search_api_call}{}}}
\subsection{Method \code{search_api_call()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baato_API$search_api_call(q, limit)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{q}}{}

\item{\code{limit}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the API response in JSON format
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{baato$search_api_call(q="shemrock nepalgunj",limit=4)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baato_API-places_api_call"></a>}}
\if{latex}{\out{\hypertarget{method-baato_API-places_api_call}{}}}
\subsection{Method \code{places_api_call()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baato_API$places_api_call(placeId)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{placeId}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the API response in JSON format
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{baato$places_api_call(placeId = 102235)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baato_API-nearby_api_call"></a>}}
\if{latex}{\out{\hypertarget{method-baato_API-nearby_api_call}{}}}
\subsection{Method \code{nearby_api_call()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baato_API$nearby_api_call(type, lat, lon, limit)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{}

\item{\code{lat}}{}

\item{\code{lon}}{}

\item{\code{limit}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the API response in JSON format
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{baato$nearby_api_call(type="school",lat=27.71765,lon=85.32691,limit=20)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baato_API-styles_api_call"></a>}}
\if{latex}{\out{\hypertarget{method-baato_API-styles_api_call}{}}}
\subsection{Method \code{styles_api_call()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baato_API$styles_api_call(style_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{style_name}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the API response in JSON format
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{baato$styles_api_call(style_name="monochrome")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baato_API-baato_directions_api"></a>}}
\if{latex}{\out{\hypertarget{method-baato_API-baato_directions_api}{}}}
\subsection{Method \code{baato_directions_api()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baato_API$baato_directions_api(points, mode)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{points}}{}

\item{\code{mode}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the API response in JSON format
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{baato$baato_directions_api(points = point, mode="car")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baato_API-clone"></a>}}
\if{latex}{\out{\hypertarget{method-baato_API-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baato_API$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
